Добавила обертку для списка фильмов и кнопки
